% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_basket_shot_charts.R
\name{gmm_basket_shot_chart}
\alias{gmm_basket_shot_chart}
\title{Gaussian mixture model for model-based shot density chart}
\usage{
gmm_basket_shot_chart(
  data,
  G = 1:9,
  modelNames = "VVV",
  prior = priorControl("defaultPrior"),
  lambda = c(0, 0),
  nstart = 100,
  ...
)
}
\arguments{
\item{data}{TODO}

\item{G}{TODO}

\item{modelNames}{TODO}

\item{lambda}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
