% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_basket_shot_charts.R
\name{gmm_basket_shot_chart_calibration}
\alias{gmm_basket_shot_chart_calibration}
\title{TODO ...}
\usage{
gmm_basket_shot_chart_calibration(
  x,
  y,
  palette = basket_offensive_areas_theme,
  ...
)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{additional arguments to be passed to the low level functions.}
}
\value{
A list of elements:
}
\description{
TODO ...
}
\examples{
TODO

}
\seealso{
\code{\link[=gmm_basket_shot_chart(), gmm_basket_shot_chart_predict]{gmm_basket_shot_chart(), gmm_basket_shot_chart_predict()}}
}
