% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_basket_shot_charts.R
\name{gmm_basket_shot_chart_predict}
\alias{gmm_basket_shot_chart_predict}
\title{Predictions for mixture-based shot density charts}
\usage{
gmm_basket_shot_chart_predict(x, y, newdata, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{newdata}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
