% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMBasketShotCharts-package.R
\docType{package}
\name{GMMBasketShotCharts-package}
\alias{GMMBasketShotCharts}
\alias{GMMBasketShotCharts-package}
\title{Gaussian mixture model for basketball shot charts estimation}
\description{
An R package implementing a model-based approach using Gaussian mixtures
for bounded data to estimate and visualize basketball shot charts
(Scrucca & Karlis, 2024).
}
\references{
Scrucca L., Karlis D. (2024) A model-based approach to shot charts estimation
in basketball. Under review.
arXiv pre-print available at https://arxiv.org/abs/2405.01182
}
\seealso{
\code{\link[=densityMclustBounded]{densityMclustBounded()}} for density estimation of bounded data.
}
\author{
\strong{Maintainer}: Luca Scrucca \email{luca.scrucca@unibo.it} (\href{https://orcid.org/0000-0003-3826-0484}{ORCID})

}
\keyword{internal}
