% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_basket_shot_charts.R
\name{plot.GMMBasketShotChart}
\alias{plot.GMMBasketShotChart}
\title{ggplot method for ... TODO}
\usage{
\method{plot}{GMMBasketShotChart}(
  x,
  prob = c(0.25, 0.5, 0.75, 0.9),
  ngrid = 100,
  theme = c("light", "dark"),
  palette = "YlGnBu",
  addpoints = FALSE,
  ...
)
}
\arguments{
\item{prob}{TODO}

\item{ngrid}{TODO}

\item{theme}{TODO}

\item{palette}{TODO}

\item{addpoints}{TODO}

\item{...}{additional arguments to be passed to the low level functions.}

\item{object}{an object of class \code{GMMBasketShotChart}}
}
\value{
A ggplot object.
}
\description{
Plots based on ggplot2 for ... TODO
}
\examples{
TODO

}
\seealso{
\code{\link[=gmm_basket_shot_chart]{gmm_basket_shot_chart()}}
}
