% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basket_court.R
\docType{data}
\name{basket_court_theme}
\alias{basket_court_theme}
\alias{basket_offensive_areas_theme}
\alias{plot_basket_court}
\alias{basket_court_offensive_areas}
\alias{basket_court_point_area}
\alias{three_point_shot}
\title{Plot basket half-court}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 3.
}
\usage{
basket_court_theme

basket_offensive_areas_theme

plot_basket_court(
  theme = c("light", "dark"),
  add = FALSE,
  use_short_three = FALSE
)

basket_court_offensive_areas(use_short_three = FALSE, ...)

basket_court_point_area(point, ...)

three_point_shot(point, ...)
}
\arguments{
\item{theme}{TODO}

\item{add}{TODO}

\item{use_short_three}{TODO}
}
\value{
A ggplot object.
}
\description{
TODO
}
\examples{
plot_basket_court()
plot_basket_court(theme = "dark")

basket_court_theme
basket_offensive_areas_theme

}
\keyword{datasets}
