% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basket_court.R
\name{plot_basket_court_offensive_areas}
\alias{plot_basket_court_offensive_areas}
\title{Plot basket half-court offensive areas}
\usage{
plot_basket_court_offensive_areas(
  theme = "light",
  use_short_three = FALSE,
  palette = basket_offensive_areas_theme,
  ...
)
}
\arguments{
\item{theme}{TODO}

\item{use_short_three}{TODO}

\item{add}{TODO}
}
\value{
A ggplot object.
}
\description{
TODO
}
\examples{
plot_basket_court_offensive_areas()
plot_basket_court_offensive_areas(theme = "dark")

}
