% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMBasketShotCharts-package.R
\docType{data}
\name{stephen_curry}
\alias{stephen_curry}
\title{Stephen Curry data}
\format{
A list of two components:
\describe{
\item{data}{a \code{data.table} containing player's features.}
\item{player}{a \code{data.table} containing player's info.}
}
}
\source{
Data obtained using the R package hoopR (Gilani, 2023) from ESPN analytics
https://www.espn.com/nba/.
}
\description{
Shooting data for Stephen Curry, 2022-23 Regular Season.
}
\keyword{datasets}
